+++ Sources/PippinLibrary/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Added
+
+- Different ways to count things in Collections: `count(where:)` and `counts`.
+
 ## [2.6.0] 2020-10-24
 
 ### Added